requirements.txt
streamlit==1.41.1
anthropic==0.45.0
openai==1.61.0
numpy>=1.24.0
tiktoken==0.8.0
python-dotenv==1.0.1
